Original 04.04.26/66- Migrar do banco de dados SQLite para o PostigreSQL/requirements.txt
streamlit>=1.40.0
sqlalchemy>=2.0.36
psycopg2-binary>=2.9.9
pandas>=2.2.0
openpyxl>=3.1.0
reportlab>=4.0.0
python-dotenv>=1.0.0
